requirements.txt
Django==2.2.4
django-phone-field==1.8.0
djangorestframework==3.10.2
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
